refactor: 코드 구조 개선 및 모듈 분리
- `KakaoLogin.js`를 `Auth.js`로 이름 변경 및 관련 함수 이동
- `Interaction.js`의 코드들을 쪼개 `User.js`, `Notification.js`, `Like.js`, `Comment.js`로 모듈 분리
- `server.js`에서 모듈별로 import 정리
- 알림 관련 API 경로를 단수형으로 변경 (`/notifications` -> `/notification`)
- 불필요한 코드 및 중복된 함수 제거
